nir: give reachable_exprs one home in the visitor module
`optimize::value_copy_demote` had its own `reachable_exprs`, hand-walking the
node graph with a stack while `niri` did the same job through `NirRefVisitor`.
Neither belongs to the module it sat in: enumerating the reachable expressions
of a body is what the visitor module is for, and `wir_build` was already
reaching into `niri` for it.

The shared one is the visitor-driven version, so it walks in order rather than
arbitrarily — both call sites in `value_copy_demote` collect into a set or
rewrite in place, neither of which the order reaches — and it handles a body
with no block structure, where indexing `body.root` would have panicked.
